% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{calculate_metrics}
\alias{calculate_metrics}
\title{Calculate Growth and Health Indices}
\usage{
calculate_metrics(df, days)
}
\arguments{
\item{df}{Dataframe containing individual fish data.}

\item{days}{Numeric. Duration of the experiment in days.}
}
\value{
Dataframe with calculated columns (WG, SGR, K, HSI, etc.).
}
\description{
Calculate Growth and Health Indices
}
