% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fish_data.R
\docType{data}
\name{Fish_data}
\alias{Fish_data}
\title{Fish growth snippet from Katie's summer 2018 field work}
\format{length is in mm, these are all integers}
\source{
Katie Zinn
}
\usage{
Fish_data
}
\description{
Fish growth snippet from Katie's summer 2018 field work
}
\keyword{datasets}
