% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guide.R
\name{fishgrowth_guide}
\alias{fishgrowth_guide}
\title{Print the FishGrowth Workflow Guide}
\usage{
fishgrowth_guide()
}
\value{
Prints text to the console.
}
\description{
Displays a cheat sheet of the package functions in the correct order of execution.
}
