% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitAgrowth.R
\name{fitAgrowth}
\alias{fitAgrowth}
\title{Fit a Length and Likelihood-based Growth Model (LLGM)}
\usage{
fitAgrowth(
  data,
  model = "HmA",
  gpars = c(60, 0.2, 0, 3),
  lbin = 1,
  lmin = 0,
  lmax = 100,
  amax = 10
)
}
\arguments{
\item{data}{A vector of data containing the length of each individual observed.}

\item{model}{The model name to be estimated, between the "*HmA*", "*HmAG*", "*HtA*" and "*HtAG*" options.}

\item{gpars}{Vector containing the four growth parameters \eqn{L_{\infty}, K, t_{0}, \sigma_{VBGE}^2}. 
The values specified for the parameters \eqn{L_{\infty}} and \eqn{K} will be used by the *HmA* and *HtA* models, 
while the *HmAG* and *HtAG* models will not use this information, since they will estimate values for 
these two parameters. The parameters \eqn{t_{0}} and \eqn{\sigma_{VBGE}^2} should be specified as 
assumptions for all models.}

\item{lbin}{Length classes amplitude.}

\item{lmin}{Lower limit of the first length class.}

\item{lmax}{Upper limit of the last length class.}

\item{amax}{Maximum age to be considered in the model.}
}
\value{
Return a fit object.
}
\description{
The fit function, fitAgrowth, used to fit the models to the length-frequency data.
}
\examples{
Pkl(pars=c(60, 0.3), gpars = c(0, 3))

}
