% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_analysis_template}
\alias{get_analysis_template}
\title{Get Master Analysis Script}
\usage{
get_analysis_template(filename = "Analysis_Workflow.R")
}
\arguments{
\item{filename}{String. Name of the file to save (default: "Analysis_Workflow.R").}
}
\value{
NULL. Saves a file to the working directory.
}
\description{
Copies the standard analysis script to your current working directory.
Use this script to run the full analysis workflow.
}
