% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_data_template}
\alias{get_data_template}
\title{Get Standard Data Template}
\usage{
get_data_template(filename = "data_template.csv")
}
\arguments{
\item{filename}{String. The name of the file to save (default: "data_template.csv").}
}
\value{
Invisible. Copies the file.
}
\description{
Copies a blank CSV template to the user's working directory.
The user can then fill this file with their experimental data.
}
