% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nllk_HmA.R
\name{nllk.HmA}
\alias{nllk.HmA}
\title{Negative logarithm likelihood of the Homoscedastic Age Model (HmA)}
\usage{
nllk.HmA(pars)
}
\arguments{
\item{pars}{Vector of the parameters to be estimated (sigma2, A), where A is a age composition vector.}
}
\value{
Return a negative logarithm likelihood of the Homoscedastic Age Model (HmA).
}
\description{
This function computes the Negative logarithm likelihood of the Homoscedastic Age Model (HmA). 
This functions is used in the \code{\link{fitAgrowth}} estimation function. 
This objective functions present a single argument, which is the vector of parameters to be estimated (pars), to enable parallel computing.
}
