% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nllk_HmAG.R, R/nllk_HtAG.R
\name{nllk.HmAG}
\alias{nllk.HmAG}
\title{Negative logarithm likelihood of the Homoscedastic Age and Growth Model (HtAG)}
\usage{
nllk.HmAG(pars)

nllk.HmAG(pars)
}
\arguments{
\item{pars}{Vector of the parameters to be estimated (q, Linf, K, A), where A is a age composition vector.}
}
\value{
Return a negative logarithm likelihood of the Homoscedastic Age and Growth Model.

Return a negative logarithm likelihood of the Heteroscedastic Age and Growth Model.
}
\description{
This function computes the Negative logarithm likelihood of the Homoscedastic Age and Growth Model (HtAG). 
This functions is used in the \code{\link{fitAgrowth}} estimation function. 
This objective functions present a single argument, which is the vector of parameters to be estimated (pars), to enable parallel computing.

This function computes the Negative logarithm likelihood of the Heteroscedastic Age and Growth Model (HtAG). 
This functions is used in the \code{\link{fitAgrowth}} estimation function. 
This objective functions present a single argument, which is the vector of parameters to be estimated (pars), to enable parallel computing.
}
