% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nllk_HtA.R
\name{nllk.HtA}
\alias{nllk.HtA}
\title{Negative logarithm likelihood of the Heteroscedastic Age Model (HtA)}
\usage{
nllk.HtA(pars)
}
\arguments{
\item{pars}{Vector of the parameters to be estimated (q, A), where A is a age composition vector.}
}
\value{
Return a negative logarithm likelihood of the Heteroscedastic Age Model.
}
\description{
This function computes the Negative logarithm likelihood of the Heteroscedastic Age Model (HtA). 
This functions is used in the \code{\link{fitAgrowth}} estimation function. 
This objective functions present a single argument, which is the vector of parameters to be estimated (pars), to enable parallel computing.
}
