% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otoliths.skj.R
\docType{data}
\name{otoliths.skj}
\alias{otoliths.skj}
\title{Otolith Data}
\format{
Data frame containing two columns:
\tabular{ll}{
  \code{age} \tab age (years)\cr
  \code{len} \tab length (cm)
}
}
\source{
Macdonald, J., Day, J., Magnusson, A., Maunder, M., Aoki, Y., Matsubara, N.,
Tsuda, Y., McKechnie, S., Teears, T., Leroy, B., Castillo-Jordán, C.,
Hampton, J., and Hamer, P. (2022).
\emph{Review and new analyses of skipjack growth in the Western and Central
Pacific Ocean}.
Western and Central Pacific Fisheries Commission Report
WCPFC-SC18-2022/SA-IP-06.
\url{https://meetings.wcpfc.int/node/16254}.
}
\usage{
otoliths.skj
}
\description{
Simulated otolith data, resembling the skipjack tuna dataset analyzed by
Macdonald et al. (2022).
}
\details{
The simulation code that was used to produce this dataset is included in the
package:
\preformatted{file.show(system.file(package="tao", "sim/simulate.R"))}
}
\examples{
otoliths.skj
}
\seealso{
\code{\link{tags.skj}} is a tagging example dataset.

\code{\link{vonbert}} fits a von Bertalanffy model.

\code{\link{tao-package}} gives an overview of the package.
}
