% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otoliths_ex.R
\docType{data}
\name{otoliths_ex}
\alias{otoliths_ex}
\title{Otolith Data}
\format{
Data frame containing two columns:
\tabular{ll}{
  \code{age} \tab age (years)\cr
  \code{len} \tab length (cm)
}
}
\source{
Macdonald, J., Day, J., Magnusson, A., Maunder, M., Aoki, Y., Matsubara, N.,
Tsuda, Y., McKechnie, S., Teears, T., Leroy, B., Castillo-Jordán, C.,
Hampton, J., and Hamer, P. (2022).
\emph{Review and new analyses of skipjack growth in the Western and Central
Pacific Ocean}.
Western and Central Pacific Fisheries Commission Report
WCPFC-SC18-2022/SA-IP-06.
\url{https://meetings.wcpfc.int/node/16254}.
}
\usage{
otoliths_ex
}
\description{
Simulated otolith data, loosely based on a skipjack tuna dataset analyzed by
Macdonald et al. (2022).
}
\details{
The simulation code that was used to produce this dataset is included in the
package:
\preformatted{file.show(system.file(package="tao", "sim/simulate.R"))}
}
\examples{
otoliths_ex
}
\seealso{
\code{otoliths_ex} and \code{\link{tags_ex}} are example datasets.

\code{\link{gcm}}, \code{\link{gompertz}}, \code{\link{richards}},
\code{\link{schnute3}}, and \code{\link{vonbert}} are alternative growth
models.

\code{\link{tao-package}} gives an overview of the package.
}
