% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otoliths_had.R
\docType{data}
\name{otoliths_had}
\alias{otoliths_had}
\title{Otolith Data (Haddock)}
\format{
Data frame containing two columns:
\tabular{ll}{
  \code{age} \tab age (years)\cr
  \code{len} \tab length (cm)
}
}
\source{
Data contributed by the \href{https://www.hafogvatn.is/en/}{Icelandic Marine
and Freshwater Research Institute}.
}
\usage{
otoliths_had
}
\description{
Otolith data for Icelandic haddock from the Icelandic spring bottom trawl
survey 2011-2020.
}
\examples{
head(otoliths_had)
}
\seealso{
\code{\link{gcm}}, \code{\link{gompertz}}/\code{\link{gompertzo}},
\code{\link{richards}}/\code{\link{richardso}}, \code{\link{schnute3}}, and
\code{\link{vonbert}}/\code{\link{vonberto}} are alternative growth models.

\code{otoliths_had}, \code{\link{otoliths_skj}}, and \code{\link{tags_skj}}
are example datasets.

\code{\link{fishgrowth-package}} gives an overview of the package.
}
