% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pkl.R
\name{Pkl}
\alias{Pkl}
\title{Transition (P) matrix}
\usage{
Pkl(pars, gpars, lmin = 0, lmax = 100, lbin = 1, amax = 10)
}
\arguments{
\item{pars}{A vector of growth parameters (Linf and K) that can be estimates}

\item{gpars}{A vector of growth parameters (t0 and sd) that can not be estimates}

\item{lmin}{A value of minimum length of the P matrix}

\item{lmax}{A value of the maximum length of the P matrix}

\item{lbin}{A value of the length bins of the P matrix}

\item{amax}{A value of the maximum age of the P matrix}
}
\value{
Return a P matrix [rows = lengths, cols = ages].
}
\description{
This function, Pkl, computes the transition matrix, based on the growth parameters of the von Bertalanffy equation.
 This function will be automatically used within the estimation function, \code{\link{fitAgrowth}}.
}
\examples{
Pkl(pars=c(60, 0.3), gpars = c(0, 3))

}
