% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_normality}
\alias{plot_normality}
\title{Plot Normality Check (Q-Q Plot)}
\usage{
plot_normality(df, parameter)
}
\arguments{
\item{df}{Dataframe (tank level).}

\item{parameter}{String. The column name to test.}
}
\description{
Visualizes residuals to check for normality assumptions.
}
