% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_band.R
\name{pred_band}
\alias{pred_band}
\title{Prediction Band}
\usage{
pred_band(age, model, level)
}
\arguments{
\item{age}{a vector of ages to calculate the prediction band.}

\item{model}{a fitted growth model.}

\item{level}{significance level.}
}
\value{
A data frame containing five columns:
\item{age}{age}
\item{Lhat}{predicted length}
\item{sigma}{growth variability}
\item{lower}{lower limit of prediction band}
\item{upper}{upper limit of prediction band}
}
\description{
Calculate a prediction band for a fitted growth curve.
}
\examples{
# Fit a model
init <- list(log_L1=log(20), log_L2=log(70), log_k=log(0.1),
             log_sigma_min=log(3), log_sigma_max=log(3))
dat <- list(Aoto=otoliths_had$age, Loto=otoliths_had$len, t1=1, t2=10)
model <- vonbert(init, dat)
fit <- nlminb(model$par, model$fn, model$gr,
              control=list(eval.max=1e4, iter.max=1e4))

# Calculate prediction band
x <- seq(1, 18, 0.5)
pred_band(x, model)

}
\seealso{
\code{\link{gcm}}, \code{\link{gompertz}}/\code{\link{gompertzo}},
\code{\link{richards}}/\code{\link{richardso}}, \code{\link{schnute3}}, and
\code{\link{vonbert}}/\code{\link{vonberto}} are alternative growth models.

\code{\link{fishgrowth-package}} gives an overview of the package.
}
