% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richards.R
\name{richards}
\alias{richards}
\title{Richards Growth Model}
\usage{
richards(par, data)
}
\arguments{
\item{par}{is a parameter list.}

\item{data}{is a data list.}
}
\value{
TMB model object, produced by \code{\link[RTMB]{MakeADFun}}.
}
\description{
Fit a Richards growth model to tags and otoliths.
}
\details{
The \code{par} list contains the following elements:
\itemize{
  \item \code{log_L1}, predicted length at age \code{t1}
  \item \code{log_L2}, predicted length at age \code{t2}
  \item \code{log_k}, growth coefficient
  \item \code{b}, shape parameter
  \item \code{log_sigma_1}, growth variability at length \code{L_short}
  \item \code{log_sigma_2}, growth variability at length \code{L_long}
  \item \code{log_age} age at release of tagged individuals (vector)
}

The \code{data} list contains the following elements:
\itemize{
  \item \code{Lrel}, length at release of tagged individuals (vector)
  \item \code{Lrec}, length at recapture of tagged individuals (vector)
  \item \code{liberty}, time at liberty of tagged individuals in years
        (vector)
  \item \code{Aoto}, age from otoliths (vector)
  \item \code{Loto}, length from otoliths (vector)
  \item \code{t1}, age where predicted length is \code{L1}
  \item \code{t2}, age where predicted length is \code{L2}
  \item \code{L_short}, length where sd(length) is \code{sigma_1}
  \item \code{L_long}, length where sd(length) is \code{sigma_2}
}
}
\note{
The Richards (1959) growth model, as parametrized by Schnute (1981, Eq. 15),
predicts length at age as:

\deqn{\hat L_t ~=~ \left[\:L_1^b\;+\;(L_2^b-L_1^b)\,
      \frac{1-e^{-k(t-t_1)}}{1-e^{-k(t_2-t_1)}}\,\right]^{1/b}}

The variability of length at age increases linearly with length,

\deqn{\sigma_L ~=~ \alpha \,+\, \beta \hat L_t}

where the slope is \eqn{\beta=(\sigma_2-\sigma_1) /
(L_\mathrm{long}-L_\mathrm{short})} and the intercept is \eqn{\alpha=\sigma_1
- \beta L_\mathrm{short}}.

The negative log-likelihood is calculated by comparing the observed and
predicted lengths:
\preformatted{
  nll_Lrel <- -dnorm(Lrel, Lrel_hat, sigma_Lrel, TRUE)
  nll_Lrec <- -dnorm(Lrec, Lrec_hat, sigma_Lrec, TRUE)
  nll_Loto <- -dnorm(Loto, Loto_hat, sigma_Loto, TRUE)
  nll <- sum(nll_Lrel) + sum(nll_Lrec) + sum(nll_Loto)
}
}
\references{
Richards, F.J. (1959).
A flexible growth function for empirical use.
\emph{Journal of Experimental Botany}, \bold{10}, 290-300.
\url{https://www.jstor.org/stable/23686557}.

Schnute, J. (1981).
A versatile growth model with statistically stable parameters.
\emph{Canadian Journal of Fisheries and Aquatic Science}, \bold{38},
1128-1140.
\doi{10.1139/f81-153}.
}
\seealso{
\code{\link{vonbert}} is another growth model.

\code{\link{otoliths_ex}} and \code{\link{tags_ex}} are example datasets.

\code{\link{tao-package}} gives an overview of the package.
}
