% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richardso.R
\name{richardso}
\alias{richardso}
\alias{richardso_curve}
\alias{richardso_objfun}
\title{Richards Growth Model (Old Style)}
\usage{
richardso(par, data, silent = TRUE, ...)

richardso_curve(t, Linf, k, tau, b)

richardso_objfun(par, data)
}
\arguments{
\item{par}{is a parameter list.}

\item{data}{is a data list.}

\item{silent}{passed to \code{\link[RTMB]{MakeADFun}}.}

\item{\dots}{passed to \code{\link[RTMB]{MakeADFun}}.}

\item{t}{age (vector).}

\item{Linf}{asymptotic maximum length.}

\item{k}{growth coefficient.}

\item{tau}{location parameter.}

\item{b}{shape parameter.}
}
\value{
The \code{richardso} function returns a TMB model object, produced by
\code{\link[RTMB]{MakeADFun}}.

The \code{richardso_curve} function returns a numeric vector of predicted
length at age.

The \code{richardso_objfun} function returns the negative log-likelihood as a
single number, describing the goodness of fit of \code{par} to the
\code{data}.
}
\description{
Fit a Richards growth model to otoliths and/or tags, using a traditional
parametrization.
}
\details{
The main function \code{richardso} creates a model object, ready for
parameter estimation. The auxiliary functions \code{richardso_curve} and
\code{richardso_objfun} are called by the main function to calculate the
regression curve and objective function value. The user can also call the
auxiliary functions directly for plotting and model exploration.

The \code{par} list contains the following elements:
\itemize{
  \item \code{log_Linf}, asymptotic maximum length
  \item \code{log_k}, growth coefficient
  \item \code{tau}, location parameter
  \item \code{b}, shape parameter
  \item \code{log_sigma_1}, growth variability at length \code{Lshort}
  \item \code{log_sigma_2} (*), growth variability at length \code{Llong}
  \item \code{log_age} (*), age at release of tagged individuals (vector)
}

*: The parameter \code{log_sigma_2} can be omitted to estimate growth
variability that does not vary with length. The parameter vector
\code{log_age} can be omitted to fit to otoliths only.

The \code{data} list contains the following elements:
\itemize{
  \item \code{Aoto} (*), age from otoliths (vector)
  \item \code{Loto} (*), length from otoliths (vector)
  \item \code{Lrel} (*), length at release of tagged individuals (vector)
  \item \code{Lrec} (*), length at recapture of tagged individuals (vector)
  \item \code{liberty} (*), time at liberty of tagged individuals in years
        (vector)
  \item \code{Lshort} (*), length where sd(length) is \code{sigma_1}
  \item \code{Llong} (*), length where sd(length) is \code{sigma_2}
}

*: The data vectors \code{Aoto} and \code{Loto} can be omitted to fit to
tagging data only. The data vectors \code{Lrel}, \code{Lrec}, and
\code{liberty} can be omitted to fit to otoliths only. The reference lengths
\code{Lshort} and \code{Llong} are only required when estimating growth
variability that varies with length, i.e., when the \code{log_sigma_2}
parameter is specified.
}
\note{
The Schnute parametrization used in \code{\link{richards}} reduces parameter
correlation and improves convergence reliability compared to the traditional
parametrization used in \code{richardso}. Therefore, the
\code{richards} parametrization can be recommended for general usage, as both
parametrizations produce the same growth curve. However, there can be some
use cases where the traditional parametrization (\code{Linf}, \code{k},
\code{tau}, \code{b}) is preferred over the Schnute parametrization
(\code{L1}, \code{L2}, \code{k}, \code{b}).

The Richards (1959) growth model, as parametrized by Tjørve and Tjørve (2010,
Eq. 4), predicts length at age as:

\deqn{\hat L_t ~=~ L_\infty\left(1\,-\,\frac{1}{b}\,
      e^{-k(t-\tau)}\right)^{\!b}}{Linf * (1 - (1/b) * exp(-k*(t-tau)))^b}

The variability of length at age increases linearly with length,

\deqn{\sigma_L ~=~ \alpha \,+\, \beta \hat L}{
      sigma_L = alpha + beta * Lhat}

where the slope is \eqn{\beta=(\sigma_2-\sigma_1) /
(L_\mathrm{long}-L_\mathrm{short})}{beta = (sigma_2-sigma_1) /
(Llong-Lshort)} and the intercept is \eqn{\alpha=\sigma_1 - \beta
L_\mathrm{short}}{alpha = sigma_1 - beta * Lshort}. Alternatively, growth
variability can be modelled as a constant
(\eqn{\sigma_L=\sigma_1}{sigma_L=sigma_1}) that does not vary with length,
see \code{log_sigma_2} above.

The negative log-likelihood is calculated by comparing the observed and
predicted lengths:
\preformatted{
  nll_Loto <- -dnorm(Loto, Loto_hat, sigma_Loto, TRUE)
  nll_Lrel <- -dnorm(Lrel, Lrel_hat, sigma_Lrel, TRUE)
  nll_Lrec <- -dnorm(Lrec, Lrec_hat, sigma_Lrec, TRUE)
  nll <- sum(nll_Loto) + sum(nll_Lrel) + sum(nll_Lrec)
}
}
\examples{
# Explore initial parameter values
plot(len~age, otoliths_skj, xlim=c(0,4), ylim=c(0,100))
x <- seq(0, 4, 0.1)
points(lenRel~I(lenRel/60), tags_skj, col=4)
points(lenRec~I(lenRel/60+liberty), tags_skj, col=3)
lines(x, richardso_curve(x, Linf=80, k=0.8, tau=-0.5, b=1), lty=2)

# Prepare parameters and data
init <- list(log_Linf=log(80), log_k=log(0.8), tau=-0.5, b=1,
             log_sigma_1=log(1), log_sigma_2=log(1),
             log_age=log(tags_skj$lenRel/60))
dat <- list(Aoto=otoliths_skj$age, Loto=otoliths_skj$len,
            Lrel=tags_skj$lenRel, Lrec=tags_skj$lenRec,
            liberty=tags_skj$liberty, Lshort=30, Llong=60)
richardso_objfun(init, dat)

# Fit model
model <- richardso(init, dat)
fit <- nlminb(model$par, model$fn, model$gr,
              control=list(eval.max=1e4, iter.max=1e4))
report <- model$report()
sdreport <- sdreport(model)

# Plot results
plot(len~age, otoliths_skj, xlim=c(0,4), ylim=c(0,100))
points(report$age, report$Lrel, col=4)
points(report$age+report$liberty, report$Lrec, col=3)
Lhat <- with(report, richardso_curve(x, Linf, k, tau, b))
lines(x, Lhat, lwd=2)

# Model summary
est <- report[c("Linf", "k", "tau", "b", "sigma_1", "sigma_2")]
est
fit[-1]
head(summary(sdreport), 6)

#############################################################################

# Fit to otoliths only
init_oto <- list(log_Linf=log(80), log_k=log(0.8), tau=-0.5, b=1,
                 log_sigma_1=log(1), log_sigma_2=log(1))
dat_oto <- list(Aoto=otoliths_skj$age, Loto=otoliths_skj$len,
                Lshort=30, Llong=60)
model_oto <- richardso(init_oto, dat_oto)
fit_oto <- nlminb(model_oto$par, model_oto$fn, model_oto$gr,
                  control=list(eval.max=1e4, iter.max=1e4))
model_oto$report()[c("Linf", "k", "tau", "b", "sigma_1", "sigma_2")]

# Same, but now estimating constant sigma instead of sigma varying by length
# We do this by omitting log_sigma_2, Lshort, Llong
init_oto_c <- list(log_Linf=log(80), log_k=log(0.8), tau=-0.5, b=1,
                   log_sigma_1=log(1))
dat_oto_c <- list(Aoto=otoliths_skj$age, Loto=otoliths_skj$len)
model_oto_c <- richardso(init_oto_c, dat_oto_c)
fit_oto_c <- nlminb(model_oto_c$par, model_oto_c$fn, model_oto_c$gr,
                    control=list(eval.max=1e4, iter.max=1e4))
model_oto_c$report()[c("Linf", "k", "tau", "b", "sigma_1")]

#############################################################################

# Fit to tags only
init_tags <- list(log_Linf=log(80), log_k=log(0.8), tau=-0.5, b=1,
                  log_sigma_1=log(1), log_sigma_2=log(1),
                  log_age=log(tags_skj$lenRel/60))
dat_tags <- list(Lrel=tags_skj$lenRel, Lrec=tags_skj$lenRec,
                 liberty=tags_skj$liberty, Lshort=30, Llong=60)
model_tags <- richardso(init_tags, dat_tags)
fit_tags <- nlminb(model_tags$par, model_tags$fn, model_tags$gr,
                   control=list(eval.max=1e4, iter.max=1e4))
model_tags$report()[c("Linf", "k", "tau", "b")]

}
\references{
Richards, F.J. (1959).
A flexible growth function for empirical use.
\emph{Journal of Experimental Botany}, \bold{10}, 290-300.
\url{https://www.jstor.org/stable/23686557}.

Tjørve, E. and Tjørve, K.M.C. (2010).
A unified approach to the Richards-model family for use in growth analyses:
Why we need only two model forms.
\emph{Journal of Theoretical Biology}, \bold{267}, 417-425.
}
\seealso{
\code{\link{gcm}}, \code{\link{gompertz}}/\code{\link{gompertzo}},
\code{\link{richards}}/\code{richardso}, \code{\link{schnute3}}, and
\code{\link{vonbert}}/\code{\link{vonberto}} are alternative growth models.

\code{\link{otoliths_had}}, \code{\link{otoliths_skj}}, and
\code{\link{tags_skj}} are example datasets.

\code{\link{fishgrowth-package}} gives an overview of the package.
}
