% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{run_stats_smart}
\alias{run_stats_smart}
\title{Run ANOVA and Post-Hoc Analysis}
\usage{
run_stats_smart(df, parameter, post_hoc = "duncan")
}
\arguments{
\item{df}{Dataframe (tank level).}

\item{parameter}{String. Name of the column to analyze.}

\item{post_hoc}{String. "tukey" or "duncan".}
}
\description{
Checks assumptions and runs appropriate stats.
}
