% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tao-package.R
\docType{package}
\name{tao-package}
\alias{tao-package}
\alias{tao}
\title{Growth from Tags and Otoliths}
\description{
Fit growth models to tagging data and otolith data, using RTMB and maximum
likelihood. The tagging data provide information about the observed length at
release and length at recapture at a later time, where the age at release is
unknown and estimated as a vector of parameters. The otoliths provide direct
observed coordinates of age and length.
}
\details{
\emph{Growth models:}
\tabular{ll}{
  \code{\link{vonbert}} \tab von Bertalanffy
}
\emph{Example data:}
\tabular{ll}{
  \code{\link{otoliths_ex}} \tab otolith data\cr
  \code{\link{tags_ex}}     \tab tagging data
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arni-magnusson/tao}
}

}
\author{
Arni Magnusson and Mark Maunder.
}
