% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonbert.R
\name{vonbert}
\alias{vonbert}
\title{von Bertalanffy}
\usage{
vonbert(par, data)
}
\arguments{
\item{par}{is a parameter list.}

\item{data}{is a data list.}
}
\value{
List produced by \code{\link[RTMB]{MakeADFun}}.
}
\description{
Fit a von Bertalanffy growth model to tags and otoliths.
}
\details{
The parameter list contains the following elements:
\itemize{
  \item \code{log_L1}
  \item \code{log_L2}
  \item \code{log_k}
  \item \code{log_sigma_1}
  \item \code{log_sigma_2}
  \item \code{log_age} (vector)
}

The data list contains the following elements:
\itemize{
  \item \code{Lrel} (vector)
  \item \code{Lrec} (vector)
  \item \code{liberty} (vector)
  \item \code{Aoto} (vector)
  \item \code{Loto} (vector)
  \item \code{L_short}
  \item \code{L_long}
}
}
