% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishGrowth-package.R
\docType{package}
\name{FishGrowth-package}
\alias{FishGrowth}
\alias{FishGrowth-package}
\title{FishGrowth: Instantaneous Growth of Fish}
\description{
This package will help you calculate the instantaneous growth of fish.     The calculation is as so: Instantaneous growth = ( (LN(final length) - LN(initial length)) / days between measurements ) * 100
}
\author{
\strong{Maintainer}: Katie Zinn \email{zinn@zoology.ubc.ca}

}
\keyword{internal}
