% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertzo.R
\name{gompertzo}
\alias{gompertzo}
\alias{gompertzo_curve}
\alias{gompertzo_objfun}
\title{Gompertz Growth Model (Old Style)}
\usage{
gompertzo(par, data, silent = TRUE, ...)

gompertzo_curve(t, Linf, k, tau)

gompertzo_objfun(par, data)
}
\arguments{
\item{par}{a parameter list.}

\item{data}{a data list.}

\item{silent}{passed to \code{\link[RTMB]{MakeADFun}}.}

\item{\dots}{passed to \code{\link[RTMB]{MakeADFun}}.}

\item{t}{age (vector).}

\item{Linf}{asymptotic maximum length.}

\item{k}{growth coefficient.}

\item{tau}{location parameter.}
}
\value{
The \code{gompertzo} function returns a TMB model object, produced by
\code{\link[RTMB]{MakeADFun}}.

The \code{gompertzo_curve} function returns a numeric vector of predicted
length at age.

The \code{gompertzo_objfun} function returns the negative log-likelihood as a
single number, describing the goodness of fit of \code{par} to the
\code{data}.
}
\description{
Fit a Gompertz growth model to otoliths and/or tags, using a traditional
parametrization.
}
\details{
The main function \code{gompertzo} creates a model object, ready for
parameter estimation. The auxiliary functions \code{gompertzo_curve} and
\code{gompertzo_objfun} are called by the main function to calculate the
regression curve and objective function value. The user can also call the
auxiliary functions directly for plotting and model exploration.

The \code{par} list contains the following elements:
\itemize{
  \item \code{log_Linf}, asymptotic maximum length
  \item \code{log_k}, growth coefficient
  \item \code{tau}, location parameter
  \item \code{log_sigma_min}, growth variability at the shortest observed
        length in the data
  \item \code{log_sigma_max} (*), growth variability at the longest observed
        length in the data
  \item \code{log_age} (*), age at release of tagged individuals (vector)
}

*: The parameter \code{log_sigma_max} can be omitted to estimate growth
variability that does not vary with length. The parameter vector
\code{log_age} can be omitted to fit to otoliths only.

The \code{data} list contains the following elements:
\itemize{
  \item \code{Aoto} (*), age from otoliths (vector)
  \item \code{Loto} (*), length from otoliths (vector)
  \item \code{Lrel} (*), length at release of tagged individuals (vector)
  \item \code{Lrec} (*), length at recapture of tagged individuals (vector)
  \item \code{liberty} (*), time at liberty of tagged individuals in years
        (vector)
}

*: The data vectors \code{Aoto} and \code{Loto} can be omitted to fit to
tagging data only. The data vectors \code{Lrel}, \code{Lrec}, and
\code{liberty} can be omitted to fit to otoliths only.
}
\note{
The Schnute parametrization used in \code{\link{gompertz}} reduces parameter
correlation and improves convergence reliability compared to the traditional
parametrization used in \code{gompertzo}. Therefore, the \code{gompertz}
parametrization can be recommended for general usage, as both
parametrizations produce the same growth curve. However, there can be some
use cases where the traditional parametrization (\code{Linf}, \code{k},
\code{tau}) is preferred over the Schnute parametrization (\code{L1},
\code{L2}, \code{k}).

Gompertz is a special case of the Richards (1959) model, where \eqn{b=0}. If
the best model fit of a \code{\link{richards}} model to a particular dataset
involves a very small estimated value of \eqn{b}, then the \code{gompertz}
model offers a preferable parametrization, as it produces the same curve
using fewer parameters.

The Gompertz (1825) growth model, as parametrized by Ricker (1979, Eq. 23)
predicts length at age as:

\deqn{\hat L_t ~=~ L_\infty\exp\!\big(\!\!-\!e^{-k(t-\tau)}\big)}{
      Lt = Linf * exp(-exp(-k * (t-tau)))}

The variability of length at age increases linearly with length,

\deqn{\sigma_L ~=~ \alpha \,+\, \beta \hat L}{
      sigma_L = alpha + beta * Lhat}

where the slope is \eqn{\beta=(\sigma_{\max}-\sigma_{\min}) /
(L_{\max}-L_{\min})}{beta = (sigma_max-sigma_min) / (L_max-L_min)}, the
intercept is \eqn{\alpha=\sigma_{\min} - \beta L_{\min}}{alpha = sigma_min -
beta * L_min}, and \eqn{L_{\min}}{L_min} and \eqn{L_{\max}}{L_max} are the
shortest and longest observed lengths in the data. Alternatively, growth
variability can be modelled as a constant
\eqn{\sigma_L=\sigma_{\min}}{sigma_L=sigma_min} that does not vary with
length, by omitting \code{log_sigma_max} from the parameter list (see above).

The negative log-likelihood is calculated by comparing the observed and
predicted lengths:
\preformatted{
  nll_Loto <- -dnorm(Loto, Loto_hat, sigma_Loto, TRUE)
  nll_Lrel <- -dnorm(Lrel, Lrel_hat, sigma_Lrel, TRUE)
  nll_Lrec <- -dnorm(Lrec, Lrec_hat, sigma_Lrec, TRUE)
  nll <- sum(nll_Loto) + sum(nll_Lrel) + sum(nll_Lrec)
}
}
\examples{
# Model 1: Fit to haddock otoliths

# Explore initial parameter values
plot(len~age, otoliths_had, xlim=c(0,18), ylim=c(0,105), pch=16,
     col="#0080a010")
x <- seq(1, 18, 0.1)
lines(x, gompertzo_curve(x, Linf=73, k=0.4, tau=2), lty=3)

# Prepare parameters and data
init <- list(log_Linf=log(73), log_k=log(0.4), tau=2,
             log_sigma_min=log(3), log_sigma_max=log(3))
dat <- list(Aoto=otoliths_had$age, Loto=otoliths_had$len)
gompertzo_objfun(init, dat)

# Fit model
model <- gompertzo(init, dat)
fit <- nlminb(model$par, model$fn, model$gr,
              control=list(eval.max=1e4, iter.max=1e4))
report <- model$report()
sdreport <- sdreport(model)

# Plot results
Lhat <- with(report, gompertzo_curve(x, Linf, k, tau))
lines(x, Lhat, lwd=2, col=2)
legend("bottomright", c("initial curve","model fit"), col=c(1,2), lty=c(3,1),
       lwd=c(1,2), bty="n", inset=0.02, y.intersp=1.25)

# Model summary
report[c("Linf", "k", "tau", "sigma_min", "sigma_max")]
fit[-1]
summary(sdreport)

# Plot 95\% prediction band
band <- pred_band(x, model)
areaplot::confplot(cbind(lower,upper)~age, band, xlim=c(0,18), ylim=c(0,100),
         ylab="len", col="mistyrose")
points(len~age, otoliths_had, xlim=c(0,18), ylim=c(0,100),
       pch=16, col="#0080a010")
lines(x, Lhat, lwd=2, col=2)
lines(lower~age, band, lty=1, lwd=0.5, col=2)
lines(upper~age, band, lty=1, lwd=0.5, col=2)

#############################################################################

# Model 2: Fit to skipjack otoliths and tags

# Explore initial parameter values
plot(len~age, otoliths_skj, xlim=c(0,4), ylim=c(0,100))
x <- seq(0, 4, 0.1)
points(lenRel~I(lenRel/60), tags_skj, col=4)
points(lenRec~I(lenRel/60+liberty), tags_skj, col=3)
lines(x, gompertzo_curve(x, Linf=75, k=1, tau=0), lty=2)
legend("bottomright", c("otoliths","tag releases","tac recaptures",
       "initial curve"), col=c(1,4,3,1), pch=c(1,1,1,NA), lty=c(0,0,0,2),
       lwd=c(1.2,1.2,1.2,1), bty="n", inset=0.02, y.intersp=1.25)

# Prepare parameters and data
init <- list(log_Linf=log(75), log_k=log(1), tau=0,
             log_sigma_min=log(3), log_sigma_max=log(3),
             log_age=log(tags_skj$lenRel/60))
dat <- list(Aoto=otoliths_skj$age, Loto=otoliths_skj$len,
            Lrel=tags_skj$lenRel, Lrec=tags_skj$lenRec,
            liberty=tags_skj$liberty)
gompertzo_objfun(init, dat)

# Fit model
model <- gompertzo(init, dat)
fit <- nlminb(model$par, model$fn, model$gr,
              control=list(eval.max=1e4, iter.max=1e4))
report <- model$report()
sdreport <- sdreport(model)

# Plot results
plot(len~age, otoliths_skj, xlim=c(0,4), ylim=c(0,100))
points(report$age, report$Lrel, col=4)
points(report$age+report$liberty, report$Lrec, col=3)
Lhat <- with(report, gompertzo_curve(x, Linf, k, tau))
lines(x, Lhat, lwd=2)
legend("bottomright", c("otoliths","tag releases","tac recaptures",
       "model fit"), col=c(1,4,3,1), pch=c(1,1,1,NA), lty=c(0,0,0,1),
       lwd=c(1.2,1.2,1.2,2), bty="n", inset=0.02, y.intersp=1.25)

# Model summary
report[c("Linf", "k", "tau", "sigma_min", "sigma_max")]
fit[-1]
head(summary(sdreport), 5)

#############################################################################

# Model 3: Fit to skipjack otoliths only

init <- list(log_Linf=log(75), log_k=log(1), tau=0,
             log_sigma_min=log(3), log_sigma_max=log(3))
dat <- list(Aoto=otoliths_skj$age, Loto=otoliths_skj$len)
model <- gompertzo(init, dat)
fit <- nlminb(model$par, model$fn, model$gr,
                  control=list(eval.max=1e4, iter.max=1e4))
model$report()[c("Linf", "k", "tau", "sigma_min", "sigma_max")]

#############################################################################

# Model 4: Fit to skipjack otoliths only,
# but now estimating constant sigma instead of sigma varying by length

# We do this by omitting log_sigma_max
init <- list(log_Linf=log(75), log_k=log(1), tau=0,
             log_sigma_min=log(3))
dat <- list(Aoto=otoliths_skj$age, Loto=otoliths_skj$len)
model <- gompertzo(init, dat)
fit <- nlminb(model$par, model$fn, model$gr,
              control=list(eval.max=1e4, iter.max=1e4))
model$report()[c("Linf", "k", "tau", "sigma_min")]

#############################################################################

# Model 5: Fit to skipjack tags only

init <- list(log_Linf=log(75), log_k=log(1), tau=0,
             log_sigma_min=log(3), log_sigma_max=log(3),
             log_age=log(tags_skj$lenRel/60))
dat <- list(Lrel=tags_skj$lenRel, Lrec=tags_skj$lenRec,
            liberty=tags_skj$liberty)
model <- gompertzo(init, dat)
fit <- nlminb(model$par, model$fn, model$gr,
              control=list(eval.max=1e4, iter.max=1e4))
model$report()[c("Linf", "k", "tau", "sigma_min", "sigma_max")]

}
\references{
Gompertz, B. (1825).
On the nature of the function expressive of the law of human mortality, and
on a new mode of determining the value of life contingencies.
\emph{Philosophical Transactions of the Royal Society}, \bold{115}, 513-583.

Ricker, W.E. (1979).
Growth rates and models.
In: W.S. Hoar et al. (eds.) \emph{Fish physiology 8: Bioenergetics and
growth}. New York: Academic Press, pp. 677-743.

The \code{\link{fishgrowth-package}} help page includes references describing
the parameter estimation method.
}
\seealso{
\code{\link{gcm}}, \code{\link{gompertz}}, \code{gompertzo},
\code{\link{richards}}, \code{\link{richards}}, \code{\link{schnute3}},
\code{\link{vonbert}}, and \code{\link{vonberto}} are alternative growth
models.

\code{\link{pred_band}} calculates a prediction band for a fitted growth
model.

\code{\link{otoliths_had}}, \code{\link{otoliths_skj}}, and
\code{\link{tags_skj}} are example datasets.

\code{\link{fishgrowth-package}} gives an overview of the package.
}
