% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_treatment_comparison}
\alias{plot_treatment_comparison}
\title{Plot Treatment Comparisons with Significance Letters}
\usage{
plot_treatment_comparison(df, parameter, y_label = NULL)
}
\arguments{
\item{df}{Dataframe (tank level).}

\item{parameter}{String. The column name to plot.}

\item{y_label}{String. Label for the Y-axis.}
}
\description{
Visualizes the distribution of a parameter and adds statistical significance letters.
}
