% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonbert.R
\name{vonbert}
\alias{vonbert}
\alias{vonbert_curve}
\alias{vonbert_objfun}
\title{Von Bertalanffy Growth Model}
\usage{
vonbert(par, data, silent = TRUE, ...)

vonbert_curve(t, L1, L2, k, t1, t2)

vonbert_objfun(par, data)
}
\arguments{
\item{par}{a parameter list.}

\item{data}{a data list.}

\item{silent}{passed to \code{\link[RTMB]{MakeADFun}}.}

\item{\dots}{passed to \code{\link[RTMB]{MakeADFun}}.}

\item{t}{age (vector).}

\item{L1}{predicted length at age \code{t1}.}

\item{L2}{predicted length at age \code{t2}.}

\item{k}{growth coefficient.}

\item{t1}{age where predicted length is \code{L1}.}

\item{t2}{age where predicted length is \code{L2}.}
}
\value{
The \code{vonbert} function returns a TMB model object, produced by
\code{\link[RTMB]{MakeADFun}}.

The \code{vonbert_curve} function returns a numeric vector of predicted
length at age.

The \code{vonbert_objfun} function returns the negative log-likelihood as a
single number, describing the goodness of fit of \code{par} to the
\code{data}.
}
\description{
Fit a von Bertalanffy growth model to tags and/or otoliths.
}
\details{
The main function \code{vonbert} creates a model object, ready for parameter
estimation. The auxiliary functions \code{vonbert_curve} and
\code{vonbert_objfun} are called by the main function to calculate the
regression curve and objective function value. The user can also call the
auxiliary functions directly for plotting and model exploration.

The \code{par} list contains the following elements:
\itemize{
  \item \code{log_L1}, predicted length at age \code{t1}
  \item \code{log_L2}, predicted length at age \code{t2}
  \item \code{log_k}, growth coefficient
  \item \code{log_sigma_1}, growth variability at length \code{L_short}
  \item \code{log_sigma_2}, growth variability at length \code{L_long}
  \item \code{log_age} (*), age at release of tagged individuals (vector)
}

*: The parameter vector \code{log_age} should be omitted when fitting to
otoliths only.

The \code{data} list contains the following elements:
\itemize{
  \item \code{Aoto} (*), age from otoliths (vector)
  \item \code{Loto} (*), length from otoliths (vector)
  \item \code{Lrel} (*), length at release of tagged individuals (vector)
  \item \code{Lrec} (*), length at recapture of tagged individuals (vector)
  \item \code{liberty} (*), time at liberty of tagged individuals in years
        (vector)
  \item \code{t1}, age where predicted length is \code{L1}
  \item \code{t2}, age where predicted length is \code{L2}
  \item \code{L_short}, length where sd(length) is \code{sigma_1}
  \item \code{L_long}, length where sd(length) is \code{sigma_2}
}

*: The data vectors \code{Aoto} and \code{Loto} should be omitted when
fitting to tagging data only. The data vectors \code{Lrel}, \code{Lrec}, and
\code{liberty} should be omitted when fitting to otoliths only.
}
\note{
The von Bertalanffy (1938) growth model, as parametrized by Schnute and
Fournier (1980), predicts length at age as:

\deqn{\hat L_t ~=~ L_1 \;+\; (L_2-L_1)\,
      \frac{1\,-\,e^{-k(t-t_1)}}{\,1\,-\,e^{-k(t_2-t_1)}\,}}{
      L1 + (L2-L1) * (1-exp(-k*(t-t1))) / (1-exp(-k*(t2-t1)))}

The variability of length at age increases linearly with length,

\deqn{\sigma_L ~=~ \alpha \,+\, \beta \hat L}{
      sigma_L = alpha + beta * Lhat}

where the slope is \eqn{\beta=(\sigma_2-\sigma_1) /
(L_\mathrm{long}-L_\mathrm{short})}{beta = (sigma_2-sigma_1) /
(L_long-L_short)} and the intercept is \eqn{\alpha=\sigma_1
- \beta L_\mathrm{short}}{alpha = sigma_1 - beta * L_short}.

The negative log-likelihood is calculated by comparing the observed and
predicted lengths:
\preformatted{
  nll_Loto <- -dnorm(Loto, Loto_hat, sigma_Loto, TRUE)
  nll_Lrel <- -dnorm(Lrel, Lrel_hat, sigma_Lrel, TRUE)
  nll_Lrec <- -dnorm(Lrec, Lrec_hat, sigma_Lrec, TRUE)
  nll <- sum(nll_Loto) + sum(nll_Lrel) + sum(nll_Lrec)
}
}
\examples{
# Explore initial parameter values
plot(len~age, otoliths_ex, xlim=c(0,4), ylim=c(0,100))
x <- seq(0, 4, 0.1)
points(lenRel~I(lenRel/60), tags_ex, col=4)
points(lenRec~I(lenRel/60+liberty), tags_ex, col=3)
lines(x, vonbert_curve(x, L1=25, L2=75, k=0.8, t1=0, t2=4), lty=2)

# Prepare parameters and data
init <- list(log_L1=log(25), log_L2=log(75), log_k=log(0.8),
             log_sigma_1=log(1), log_sigma_2=log(1),
             log_age=log(tags_ex$lenRel/60))
dat <- list(Aoto=otoliths_ex$age, Loto=otoliths_ex$len,
            Lrel=tags_ex$lenRel, Lrec=tags_ex$lenRec,
            liberty=tags_ex$liberty, t1=0, t2=4, L_short=30, L_long=60)
vonbert_objfun(init, dat)

# Fit model
model <- vonbert(init, dat)
fit <- nlminb(model$par, model$fn, model$gr,
              control=list(eval.max=1e4, iter.max=1e4))
report <- model$report()
sdreport <- sdreport(model, getReportCovariance=FALSE)

# Plot results
plot(len~age, otoliths_ex, xlim=c(0,4), ylim=c(0,100))
points(report$age, tags_ex$lenRel, col=4)
points(report$age+tags_ex$liberty, tags_ex$lenRec, col=3)
lines(report$age_seq, report$curve, lwd=2)

# Model summary
est <- report[c("L1", "L2", "k", "sigma_1", "sigma_2")]
est
fit[-1]
head(summary(sdreport), 5)

#############################################################################

# Fit to otoliths only
init_oto <- list(log_L1=log(25), log_L2=log(75), log_k=log(0.8),
                 log_sigma_1=log(1), log_sigma_2=log(1))
dat_oto <- list(Aoto=otoliths_ex$age, Loto=otoliths_ex$len, t1=0, t2=4,
                L_short=30, L_long=60)
model_oto <- vonbert(init_oto, dat_oto)
fit_oto <- nlminb(model_oto$par, model_oto$fn, model_oto$gr,
                  control=list(eval.max=1e4, iter.max=1e4))
model_oto$report()[c("L1", "L2", "k")]

#############################################################################

# Fit to tags only
init_tags <- list(log_L1=log(25), log_L2=log(75), log_k=log(0.8),
                  log_sigma_1=log(1), log_sigma_2=log(1),
                  log_age=log(tags_ex$lenRel/60))
dat_tags <- list(Lrel=tags_ex$lenRel, Lrec=tags_ex$lenRec,
                 liberty=tags_ex$liberty, t1=0, t2=4, L_short=30, L_long=60)
model_tags <- vonbert(init_tags, dat_tags)
fit_tags <- nlminb(model_tags$par, model_tags$fn, model_tags$gr,
                   control=list(eval.max=1e4, iter.max=1e4))
model_tags$report()[c("L1", "L2", "k")]

}
\references{
von Bertalanffy, L. (1938).
A quantitative theory of organic growth.
\emph{Human Biology}, \bold{10}, 181-213.
\url{https://www.jstor.org/stable/41447359}.

Schnute, J. and Fournier, D. (1980).
A new approach to length-frequency analysis: Growth structure.
\emph{Canadian Journal of Fisheries and Aquatic Science}, \bold{37},
1337-1351.
\doi{10.1139/f80-172}.
}
\seealso{
\code{\link{gcm}}, \code{\link{gompertz}}, \code{\link{richards}},
\code{\link{schnute3}}, and \code{vonbert} are alternative growth models.

\code{\link{otoliths_ex}} and \code{\link{tags_ex}} are example datasets.

\code{\link{tao-package}} gives an overview of the package.
}
