% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vonbert.R
\name{vonbert}
\alias{vonbert}
\alias{vonbert_curve}
\alias{vonbert_objfun}
\title{Von Bertalanffy Growth Model}
\usage{
vonbert(par, data, silent = TRUE, ...)

vonbert_curve(t, L1, L2, k, t1, t2)

vonbert_objfun(par, data)
}
\arguments{
\item{par}{a parameter list.}

\item{data}{a data list.}

\item{silent}{passed to \code{\link[RTMB]{MakeADFun}}.}

\item{\dots}{passed to \code{\link[RTMB]{MakeADFun}}.}

\item{t}{age (vector).}

\item{L1}{predicted length at age \code{t1}.}

\item{L2}{predicted length at age \code{t2}.}

\item{k}{growth coefficient.}

\item{t1}{age where predicted length is \code{L1}.}

\item{t2}{age where predicted length is \code{L2}.}
}
\value{
The \code{vonbert} function returns a TMB model object, produced by
\code{\link[RTMB]{MakeADFun}}.

The \code{vonbert_curve} function returns a numeric vector of predicted
length at age.

The \code{vonbert_objfun} function returns the negative log-likelihood as a
single number, describing the goodness of fit of \code{par} to the
\code{data}.
}
\description{
Fit a von Bertalanffy growth model to tags and otoliths.
}
\details{
The main function \code{vonbert} creates a model object, ready for parameter
estimation. The auxiliary functions \code{vonbert_curve} and
\code{vonbert_objfun} are called by the main function to calculate the
regression curve and objective function value. The user can also call the
auxiliary functions directly for plotting and model exploration.

The \code{par} list contains the following elements:
\itemize{
  \item \code{log_L1}, predicted length at age \code{t1}
  \item \code{log_L2}, predicted length at age \code{t2}
  \item \code{log_k}, growth coefficient
  \item \code{log_sigma_1}, growth variability at length \code{L_short}
  \item \code{log_sigma_2}, growth variability at length \code{L_long}
  \item \code{log_age} age at release of tagged individuals (vector)
}

The \code{data} list contains the following elements:
\itemize{
  \item \code{Lrel}, length at release of tagged individuals (vector)
  \item \code{Lrec}, length at recapture of tagged individuals (vector)
  \item \code{liberty}, time at liberty of tagged individuals in years
        (vector)
  \item \code{Aoto}, age from otoliths (vector)
  \item \code{Loto}, length from otoliths (vector)
  \item \code{t1}, age where predicted length is \code{L1}
  \item \code{t2}, age where predicted length is \code{L2}
  \item \code{L_short}, length where sd(length) is \code{sigma_1}
  \item \code{L_long}, length where sd(length) is \code{sigma_2}
}
}
\note{
The von Bertalanffy (1938) growth model, as parametrized by Schnute and
Fournier (1980), predicts length at age as:

\deqn{\hat L_t ~=~ L_1 \;+\; (L_2-L_1)\,
      \frac{1\,-\,e^{-k(t-t_1)}}{\,1\,-\,e^{-k(t_2-t_1)}\,}}{
      L1 + (L2-L1) * (1-exp(-k*(t-t1))) / (1-exp(-k*(t2-t1)))}

The variability of length at age increases linearly with length,

\deqn{\sigma_L ~=~ \alpha \,+\, \beta \hat L}{
      sigma_L = alpha + beta * Lhat}

where the slope is \eqn{\beta=(\sigma_2-\sigma_1) /
(L_\mathrm{long}-L_\mathrm{short})}{beta = (sigma_2-sigma_1) /
(L_long-L_short)} and the intercept is \eqn{\alpha=\sigma_1
- \beta L_\mathrm{short}}{alpha = sigma_1 - beta * L_short}.

The negative log-likelihood is calculated by comparing the observed and
predicted lengths:
\preformatted{
  nll_Lrel <- -dnorm(Lrel, Lrel_hat, sigma_Lrel, TRUE)
  nll_Lrec <- -dnorm(Lrec, Lrec_hat, sigma_Lrec, TRUE)
  nll_Loto <- -dnorm(Loto, Loto_hat, sigma_Loto, TRUE)
  nll <- sum(nll_Lrel) + sum(nll_Lrec) + sum(nll_Loto)
}
}
\references{
von Bertalanffy, L. (1938).
A quantitative theory of organic growth.
\emph{Human Biology}, \bold{10}, 181-213.
\url{https://www.jstor.org/stable/41447359}.

Schnute, J. and Fournier, D. (1980).
A new approach to length-frequency analysis: Growth structure.
\emph{Canadian Journal of Fisheries and Aquatic Science}, \bold{37},
1337-1351.
\doi{10.1139/f80-172}.
}
\seealso{
\code{\link{gcm}}, \code{\link{richards}}, \code{\link{schnute3}}, and
\code{vonbert} are alternative growth models.

\code{\link{otoliths_ex}} and \code{\link{tags_ex}} are example datasets.

\code{\link{tao-package}} gives an overview of the package.
}
